Fix weight factors of the mass matrix.
If each of the equations:

Cj0 * bar.x + Cj1 * bar.y + Cj2 * bar.z = Si

in [Mij] [Cj] = [Si]

is weighted by w. When we take the product:

[Mji][Mij] [Cj] = [Mji][Si]

The w factors in [Mji][Mij] and [Mji][Si] should be squared, since they appear in both sides of the product. That is:

[Wi][Mij] [Cj] = [Wi][Si] => [Mji][Wi]^2[Mij] [Cj] = [Mji][Wi]^2[Si]